requirements.txt
pandas
numpy
plotly
matplotlib
seaborn
scikit-learn
jupyter
streamlit
openai
python-dotenv
groq
duckdb
tabulate
streamlit-aggrid